import seed vm to its own file
